chore: add detailed iri server logs

+++ internal/server/machine_delete.go
@@ -8,14 +8,18 @@ import (
 	"fmt"
 
 	iri "github.com/ironcore-dev/ironcore/iri/apis/machine/v1alpha1"
+	internalutils "github.com/ironcore-dev/libvirt-provider/internal/utils"
 )
 
 func (s *Server) DeleteMachine(ctx context.Context, req *iri.DeleteMachineRequest) (*iri.DeleteMachineResponse, error) {
-	log := s.loggerFrom(ctx)
+	if req == nil {
+		return nil, convertInternalErrorToGRPC(wrapErrorRequestIsNil(ErrInvalidRequest))
+	}
+	log := s.loggerFrom(ctx, internalutils.LogKeyMachineID, req.MachineId)
 
 	log.V(1).Info("Deleting machine")
 	if err := s.machineStore.Delete(ctx, req.MachineId); err != nil {
-		return nil, convertInternalErrorToGRPC(fmt.Errorf("error deleting machine: '%s': %w", req.MachineId, err))
+		return nil, convertInternalErrorToGRPC(fmt.Errorf("error deleting machine: %w", err))
 	}
 
 	return &iri.DeleteMachineResponse{}, nil
